perf(fibers): avoid heap allocation for scoped ProactorPool blocking counters
Add FiberBlockingCounter, backed by an EmbeddedBlockingCounter embedded
directly in FiberInterface instead of a heap-allocated shared_ptr. Lifetime
is anchored to the owning fiber's existing intrusive_ptr refcount, so remote
Dec()-calling closures keep the fiber (and its counter) alive through the
trailing notifyAll() without any extra allocation.

Used for ProactorPool::Await/AwaitBrief/AwaitFiberOnAll, which are tight,
single-call, scoped rendezvous points. BlockingCounter (heap/shared_ptr
based) stays as-is for long-lived owners like AcceptServer::ref_bc_, whose
Run()/Wait() split across calls would corrupt a shared per-fiber slot.
